- Refactored jet stuff to its own package - Introduced CANCELLED state for Jet jobs
