+++ base/commands/job/job_resume.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/hazelcast/hazelcast-commandline-client/clc"
 	. "github.com/hazelcast/hazelcast-commandline-client/internal/check"
+	"github.com/hazelcast/hazelcast-commandline-client/internal/jet"
 	"github.com/hazelcast/hazelcast-commandline-client/internal/plug"
 	"github.com/hazelcast/hazelcast-commandline-client/internal/proto/codec"
 )
@@ -27,14 +28,14 @@ func (cm ResumeCmd) Exec(ctx context.Context, ec plug.ExecContext) error {
 	if err != nil {
 		return err
 	}
-	jm, err := newJobNameToIDMap(ctx, ec, false)
+	jm, err := jet.NewJobNameToIDMap(ctx, ec, false)
 	if err != nil {
 		return err
 	}
 	nameOrID := ec.Args()[0]
 	jid, ok := jm.GetIDForName(nameOrID)
 	if !ok {
-		return ErrInvalidJobID
+		return jet.ErrInvalidJobID
 	}
 	_, stop, err := ec.ExecuteBlocking(ctx, func(ctx context.Context, sp clc.Spinner) (any, error) {
 		sp.SetText(fmt.Sprintf("Resuming job: %s", idToString(jid)))
@@ -51,7 +52,7 @@ func (cm ResumeCmd) Exec(ctx context.Context, ec plug.ExecContext) error {
 	if ec.Props().GetBool(flagWait) {
 		msg := fmt.Sprintf("Waiting for job %s to start", nameOrID)
 		ec.Logger().Info(msg)
-		err = WaitJobState(ctx, ec, msg, nameOrID, statusRunning, 2*time.Second)
+		err = WaitJobState(ctx, ec, msg, nameOrID, jet.JobStatusRunning, 2*time.Second)
 		if err != nil {
 			return err
 		}
